fix: check박스 단일로 나오게 수정
